requirements.txt
aiogram==3.7.0
qrcode==7.4.2
Pillow==10.3.0